fix bug header not valid

+++ api/posts.php
@@ -79,8 +79,8 @@ if ($requestMethod == 'GET' && $post->id != false) {
 }
 
 if ($requestMethod == 'POST') {
-    header('Access-Control-Allow-Methods : POST');
-    header('Access-Control-Allow-Headers : Access-Control-Allow-Headers,Content-Type,Access-Control-Allow-Methods,Authorization,X-Requested-With');
+    header('Access-Control-Allow-Methods: POST');
+    header('Access-Control-Allow-Headers: Access-Control-Allow-Headers,Content-Type,Access-Control-Allow-Methods,Authorization,X-Requested-With');
 
     // get raw data
     $data = json_decode(file_get_contents("php://input"));
@@ -89,6 +89,10 @@ if ($requestMethod == 'POST') {
     $post->body = $data->body;
     $post->author = $data->author;
     $post->categoryId = $data->category_id;
+
+    if (is_null($data)) {
+        die("Invalid JSON data.");
+    }
 
     if ($post->create()) {
         echo json_encode(
@@ -102,8 +106,8 @@ if ($requestMethod == 'POST') {
 }
 
 if ($requestMethod == 'PUT') {
-    header('Access-Control-Allow-Methods : PUT');
-    header('Access-Control-Allow-Headers : Access-Control-Allow-Headers,Content-Type,Access-Control-Allow-Methods,Authorization,X-Requested-With');
+    header('Access-Control-Allow-Methods: PUT');
+    header('Access-Control-Allow-Headers: Access-Control-Allow-Headers,Content-Type,Access-Control-Allow-Methods,Authorization,X-Requested-With');
 
     // get raw data
     $data = json_decode(file_get_contents("php://input"));
@@ -113,6 +117,10 @@ if ($requestMethod == 'PUT') {
     $post->body = $data->body;
     $post->author = $data->author;
     $post->categoryId = $data->category_id;
+
+    if (is_null($data)) {
+        die("Invalid JSON data.");
+    }
 
     if ($post->update()) {
         echo json_encode(
@@ -126,13 +134,17 @@ if ($requestMethod == 'PUT') {
 }
 
 if ($requestMethod == 'DELETE') {
-    header('Access-Control-Allow-Methods : DELETE');
-    header('Access-Control-Allow-Headers : Access-Control-Allow-Headers,Content-Type,Access-Control-Allow-Methods,Authorization,X-Requested-With');
+    header('Access-Control-Allow-Methods: DELETE');
+    header('Access-Control-Allow-Headers: Access-Control-Allow-Headers,Content-Type,Access-Control-Allow-Methods,Authorization,X-Requested-With');
 
     // get raw data
     $data = json_decode(file_get_contents("php://input"));
 
     $post->id = $data->id;
+
+    if (is_null($data)) {
+        die("Invalid JSON data.");
+    }
 
     if ($post->delete()) {
         echo json_encode(
